chore(waf): processing methods of other files used by waf dedicated domain resource
